worker: don't fail if not created image is removed

+++ worker.py
@@ -27,7 +27,8 @@ class GarbageCollector(threading.Thread):
         image_hash, image_path = image
         self.log.debug("remove outdated image %s", image)
         self.database.del_image(image_hash)
-        shutil.rmtree(image_path)
+        if os.path.exists(image_path):
+            shutil.rmtree(image_path)
 
     def run(self):
         while True:
